refactor: move calculations directory to root of library and rename to /domain
