+++ packages/oasis-actions/src/domain/Position.test.ts
@@ -1,7 +1,7 @@
 import BigNumber from 'bignumber.js'
 import { expect } from 'chai'
 
-import { ONE, ZERO } from '../constants'
+import { ONE, ZERO } from '../helpers/constants'
 import { Position } from './Position'
 import { RiskRatio } from './RiskRatio'
 import { testDataSources } from './test-scenarios/generateTestData'
