+++ packages/oasis-actions/src/strategies/aave/open/open.ts
@@ -2,9 +2,9 @@ import BigNumber from 'bignumber.js'
 import { providers } from 'ethers'
 
 import { Unbox } from '../../../../../../helpers/types/common'
+import { IBaseSimulatedTransition, Position } from '../../../domain/Position'
+import { IRiskRatio } from '../../../domain/RiskRatio'
 import { amountFromWei, amountToWei } from '../../../helpers'
-import { IBaseSimulatedTransition, Position } from '../../../helpers/calculations/Position'
-import { IRiskRatio } from '../../../helpers/calculations/RiskRatio'
 import { DEFAULT_FEE, NO_FEE, TYPICAL_PRECISION, ZERO } from '../../../helpers/constants'
 import { acceptedFeeToken } from '../../../helpers/swap/acceptedFeeToken'
 import { feeResolver } from '../../../helpers/swap/feeResolver'
