+++ packages/oasis-actions/src/types/aave/AavePosition.ts
@@ -1,7 +1,7 @@
 import BigNumber from 'bignumber.js'
 import { Optional } from 'utility-types'
 
-import { IPositionCategory, Position, PositionBalance } from '../../helpers/calculations/Position'
+import { IPositionCategory, Position, PositionBalance } from '../../domain/Position'
 import { Address } from '../StrategyParams'
 import { AAVETokens } from './tokens'
